backend/requirements.txt
# FastAPI
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# Environment
python-dotenv>=1.0.1

# AI / LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.2.0

# Vector DB + Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Data
pydantic>=2.7.0
pandas>=2.2.0
numpy>=1.26.0

# Lab parsing
pdfplumber>=0.11.0

# Optional OCR (image lab reports)
# pillow>=10.0.0
# pytesseract>=0.3.10
